% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{memcrispr.compareControlGuides}
\alias{memcrispr.compareControlGuides}
\title{Compare the counts of control guides for all samples in the experiment}
\usage{
memcrispr.compareControlGuides(countTable, controlString = NULL)
}
\arguments{
\item{countTable}{Standard \code{memcrispr} table of count data.}

\item{controlString}{Specify a string that identifys the control guides.
This is often 'control', 'non-targetting' etc.  If this is left as NULL
all guides will be compared, which can be very slow and/or crash.}
}
\description{
Compare the counts of control guides for all samples in the experiment
}
