% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats2.R
\name{memcrispr.fitModel}
\alias{memcrispr.fitModel}
\title{fitting the mixed effects model}
\usage{
memcrispr.fitModel(countsTable, controlString = "control")
}
\value{
data.table of genes with log fold-change and adjusted p.values
}
\description{
model fitting
}
