% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{memcrispr.guideDistribution}
\alias{memcrispr.guideDistribution}
\title{Plot distribution of guide counts}
\usage{
memcrispr.guideDistribution(countTable)
}
\description{
Plot distribution of guide counts
}
