% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEMcrispR.R
\docType{package}
\name{MEMcrispR}
\alias{MEMcrispR}
\alias{MEMcrispR-package}
\title{Mixed Effects Tool for Analysis of CRISPR screens}
\description{
Alignment, quality control, and analysis of CRISPR screen 
experiments.
}
