% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{metacrispr.align}
\alias{metacrispr.align}
\title{Align reads to reference}
\usage{
metacrispr.align(path, sampleSheet = NULL, outputDir = NULL,
  guideLibraries = c("GeCKOv2_A", "GeCKOv2_B"), guidePrefix = NULL,
  guideSuffix = NULL, ncores = 4)
}
\arguments{
\item{path}{Full file path to the folder containing the data.}

\item{sampleSheet}{Name of the file specifying the structure of the
experiment.  If left NULL a default of 'SampleSheet.txt' is used.}

\item{outputDir}{Full path to the location where output files should be 
written.  If left NULL the current working directory is used.}

\item{guideLibraries}{Names of libraries to align against. These can be
paths to files for custom libraries not included in the package.}

\item{guidePrefix}{Depending upon the CRISPR library you are using you may
expect all your reads to contain a common sequence before (5` of) the 
particular guide is read.  Provide this common sequence here to aid with 
alignment.}

\item{guideSuffix}{Similar to \code{guidePrefix}, but for a common sequence
that appears after (3` of) the guide sequence.}

\item{ncores}{The number of CPU cores to be used by the Rsubread 
aligner. Defaults to 4.}
}
\description{
This function expects to be provided a folder containing FASTQ files 
produced by sequencing a genetic screening experiment.  It will align
the reads against a reference sequence generated from specified guide
libraries.
}
