% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting-sample_specific.R
\name{metacrispr.fitModel.sampleSpecific}
\alias{metacrispr.fitModel.sampleSpecific}
\title{fitting the mixed effects model}
\usage{
metacrispr.fitModel.sampleSpecific(countsTable, controlString = "control")
}
\value{
data.table of genes with log fold-change and adjusted p.values
}
\description{
model fitting
}
