% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{metacrispr.guideDistribution}
\alias{metacrispr.guideDistribution}
\title{Plot distribution of guide counts}
\usage{
metacrispr.guideDistribution(countTable)
}
\description{
Plot distribution of guide counts
}
