% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/METAcrispR.R
\docType{package}
\name{METAcrispR}
\alias{METAcrispR}
\alias{METAcrispR-package}
\title{Mixed Effects Tool for Analysis of CRISPR screens}
\description{
Alignment, quality control, and analysis of CRISPR screen 
experiments.
}
