% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalizeBetweenSamples.old}
\alias{normalizeBetweenSamples.old}
\title{Normalizing between samples to account for varying total read counts}
\usage{
normalizeBetweenSamples.old(countTable)
}
\value{
data.table of counts
}
\description{
Normalization
}
