% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalizeWithinSamples}
\alias{normalizeWithinSamples}
\title{Normalize individual guides within a sample to account for GC content}
\usage{
normalizeWithinSamples(countTable)
}
\value{
data.table of counts
}
\description{
Normalization
}
