% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting-all_samples.R
\name{PanCell.fitModel}
\alias{PanCell.fitModel}
\title{fitting the mixed effects model}
\usage{
PanCell.fitModel(countsTable, controlString = "control")
}
\value{
data.table of genes with log fold-change and adjusted p.values
}
\description{
model fitting
}
