% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{screenR.count}
\alias{screenR.count}
\title{Count occurrences of each guide in aligned data.}
\usage{
screenR.count(path, sampleSheet = NULL, outputDir = NULL,
  guideLibraries = c("GeCKOv2_A", "GeCKOv2_B"), ncores = 4)
}
\arguments{
\item{path}{Full file path to the folder containing the data.}

\item{sampleSheet}{Name of the file specifying the structure of the
experiment.  If left NULL a default of 'SampleSheet.txt' is used.}

\item{outputDir}{Full path to the location where output files should be 
written.  If left NULL the current working directory is used.}

\item{guideLibraries}{Names of libraries to align against.  These can be
paths to files for custom libraries not included in the package.}

\item{ncores}{The number of CPU cores to be used by the Rsubread 
aligner. Defaults to 4.}
}
\description{
This function expects to be provided a folder containing BAM files 
produced either by \code{\link{screenR.align}} or another aligner. It will
count the number of reads mapping to each guide in the reference and 
produce a table of the same structure as produced by \code{samtools idxstat}.
}
