% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_processing.R
\name{screenR.readCounts}
\alias{screenR.readCounts}
\title{Read text files of counts}
\usage{
screenR.readCounts(path, sampleSheet = NULL, guideLibraries = NULL)
}
\arguments{
\item{path}{File path to the folder containing the data.}

\item{sampleSheet}{Name of the sample sheet file. If left as \code{NULL}
this will default to `\code{SampleSheet.txt}'.}
}
\value{
data.table of counts
}
\description{
This is the main data import function for the \code{screenR} package.  It 
expects to be provided with a directory location containing a sample sheet
and text files holding the observed counts for each guide, with one file per
sequencing library.
}
