% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{screenR.volcanoPlot}
\alias{screenR.volcanoPlot}
\title{Create volcano plot comparing fold-change and p-value for each gene}
\usage{
screenR.volcanoPlot(topTable, fc.thresh = 1, p.thresh = 0.05)
}
\arguments{
\item{topTable}{Standard \code{screenR} table of model results.  Typically
produced by \code{\link{screenR.fitModel}}.}

\item{fc.thresh}{Specify a fold change cutoff. Expects to be provide the raw
fold change, which will be automatically converted to log2 scale for 
plotting.}

\item{p.thresh}{Specify a p-value cutoff.  This exepcts to be provided with
the actual p-value e.g. 0.05, which will automatically be converted to 
-log10 scale for plotting.}
}
\description{
For every gene in a model results table, this function plots the \code{log2}
fold change against the \code{-log10} p-value.  Genes that fall below 
specified p-value threshold and above a fold change threshold are 
highlighted and named in the plot.
}
